refactor: move morphology
