--- crates/trie-morphology/src/fm/mod.rs
@@ -1,4 +1,0 @@
-mod morphology;
-
-pub use morphology::Error;
-pub use morphology::Morphology;
