+++ crates/trie-morphology/src/morphology.rs
